+++ nostr/nostr-core/src/main/java/org/tbk/nostr/util/MoreKinds.java
@@ -9,28 +9,23 @@ public final class MoreKinds {
     private static final Kind KIND_USER_METADATA = Kind.of(0);
     private static final Kind SHORT_TEXT_NOTE_KIND = Kind.of(1);
     private static final Kind KIND_FOLLOWS = Kind.of(3);
-    private static final Kind REPLACEABLE_KIND_LOWER_BOUND_INCLUSIVE = Kind.of(10_000);
-    private static final Kind REPLACEABLE_KIND_UPPER_BOUND_EXCLUSIVE = Kind.of(20_000);
     private static final Range<Kind> REPLACEABLE_KIND_RANGE = Range.closedOpen(
-            REPLACEABLE_KIND_LOWER_BOUND_INCLUSIVE,
-            REPLACEABLE_KIND_UPPER_BOUND_EXCLUSIVE
+            Kind.of(10_000),
+            Kind.of(20_000)
     );
 
     private static final Predicate<Kind> REPLACEABLE_KIND_PREDICATE = REPLACEABLE_KIND_RANGE
             .or(KIND_USER_METADATA::equals)
             .or(KIND_FOLLOWS::equals);
 
-    private static final Kind EPHEMERAL_KIND_LOWER_BOUND_INCLUSIVE = REPLACEABLE_KIND_UPPER_BOUND_EXCLUSIVE;
-    private static final Kind EPHEMERAL_KIND_UPPER_BOUND_EXCLUSIVE = Kind.of(30_000);
     private static final Range<Kind> EPHEMERAL_KIND_RANGE = Range.closedOpen(
-            EPHEMERAL_KIND_LOWER_BOUND_INCLUSIVE,
-            EPHEMERAL_KIND_UPPER_BOUND_EXCLUSIVE
+            REPLACEABLE_KIND_RANGE.upperEndpoint(),
+            Kind.of(30_000)
     );
-    private static final Kind PARAMETERIZED_REPLACEABLE_KIND_LOWER_BOUND_INCLUSIVE = EPHEMERAL_KIND_UPPER_BOUND_EXCLUSIVE;
-    private static final Kind PARAMETERIZED_REPLACEABLE_KIND_UPPER_BOUND_EXCLUSIVE = Kind.of(40_000);
-    private static final Range<Kind> PARAMETERIZED_REPLACEABLE_KIND_RANGE = Range.closedOpen(
-            PARAMETERIZED_REPLACEABLE_KIND_LOWER_BOUND_INCLUSIVE,
-            PARAMETERIZED_REPLACEABLE_KIND_UPPER_BOUND_EXCLUSIVE
+
+    private static final Range<Kind> ADDRESSABLE_KIND_RANGE = Range.closedOpen(
+            EPHEMERAL_KIND_RANGE.upperEndpoint(),
+            Kind.of(40_000)
     );
 
     private MoreKinds() {
@@ -58,8 +53,8 @@ public final class MoreKinds {
         return EPHEMERAL_KIND_RANGE;
     }
 
-    public static Range<Kind> kindParameterizedReplaceableRange() {
-        return PARAMETERIZED_REPLACEABLE_KIND_RANGE;
+    public static Range<Kind> kindAddressableRange() {
+        return ADDRESSABLE_KIND_RANGE;
     }
 
     public static boolean isReplaceable(int kind) {
@@ -92,15 +87,15 @@ public final class MoreKinds {
         }
     }
 
-    public static boolean isParameterizedReplaceable(int kind) {
-        return PARAMETERIZED_REPLACEABLE_KIND_RANGE.contains(Kind.of(kind));
+    public static boolean isAddressable(int kind) {
+        return ADDRESSABLE_KIND_RANGE.contains(Kind.of(kind));
     }
 
-    public static void checkParameterizedReplaceable(int kind) {
-        if (!isParameterizedReplaceable(kind)) {
-            throw new IllegalArgumentException("Given kind is not parameterized replaceable. Must be %d <= n < %d, got: %d".formatted(
-                    PARAMETERIZED_REPLACEABLE_KIND_RANGE.lowerEndpoint().getValue(),
-                    PARAMETERIZED_REPLACEABLE_KIND_RANGE.upperEndpoint().getValue(),
+    public static void checkAddressable(int kind) {
+        if (!isAddressable(kind)) {
+            throw new IllegalArgumentException("Given kind is not addressable. Must be %d <= n < %d, got: %d".formatted(
+                    ADDRESSABLE_KIND_RANGE.lowerEndpoint().getValue(),
+                    ADDRESSABLE_KIND_RANGE.upperEndpoint().getValue(),
                     kind
             ));
         }